requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
PyYAML>=5.4.1